+++ NEWS.md
@@ -1,5 +1,11 @@
 # sdmTMB (development version)
 
+* Allow for fitting uncorrelated random slopes and intercepts with the form
+  `(1 + slope_var || group)`, making sure that results are displayed correctly.
+
+* The correctly transformed correlation between a random slope and intercept is 
+  now displayed by `print()` and `tidy()`.
+
 * Use `tryCatch()` on Newton loops in optimization to abort those steps
   instead of producing an error.